Extracted communication interfaces to discrete script
