+++ Runtime/Scripts/Behaviors/CommunicationInterfaces.cs
@@ -1,0 +1,21 @@
+namespace BCIEssentials
+{
+    using LSLFramework;
+
+
+    public interface IHasInstanceID
+    {
+        public int GetInstanceID();
+    }
+
+
+    public interface IMarkerSource : IHasInstanceID
+    {
+        public MarkerWriter MarkerWriter { set; }
+    }
+
+    public interface IPredictionSink : IHasInstanceID
+    {
+        public void OnPrediction(Prediction prediction);
+    }
+}
