+++ Runtime/Scripts/Behaviors/CommunicationInterfaces.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: a2535f0ded32c99478d6e48d2dd21006
